third_party/WebKit/LayoutTests/platform/win/external/wpt/html/semantics/scripting-1/the-script-element/script-not-executed-after-shutdown-expected.txt

This is a testharness.js-based test.
PASS Script is not executed after script thread is shutdown 
Harness: the test ran to completion.
